backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.34.69
click==8.1.8
efficientnet_pytorch==0.7.1
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
setuptools==78.0.2
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
uvicorn==0.34.0
